pyproject.toml
[project]
name = "cute-kernels"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "nvidia-cutlass-dsl>=4.3.5",
    "torch>=2.10.0",
]
